feat(spec-062): plow through 10 deferred tasks — 95/103 done

+++ src/babylon/persistence/conservation_audit_query.py
@@ -28,8 +28,8 @@ _COUNT_BY_SEVERITY = """
 SELECT severity, COUNT(*)
 FROM conservation_audit_log
 WHERE session_id = %(session_id)s
-  AND (%(tick_lo)s IS NULL OR tick >= %(tick_lo)s)
-  AND (%(tick_hi)s IS NULL OR tick <= %(tick_hi)s)
+  AND (%(tick_lo)s::integer IS NULL OR tick >= %(tick_lo)s::integer)
+  AND (%(tick_hi)s::integer IS NULL OR tick <= %(tick_hi)s::integer)
 GROUP BY severity
 """
